Moved folder names and renamed project/package for version 2.0.0, which also removes the ISerializable interface for a better way to detect JSON-serializable value types.
