+++ BassClefStudio.NET.Serialization.Tests/BassClefStudio.NET.Serialization.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\BassClefStudio.Serialization\BassClefStudio.Serialization.csproj" />
+    <ProjectReference Include="..\BassClefStudio.NET.Serialization\BassClefStudio.NET.Serialization.csproj" />
   </ItemGroup>
 
 </Project>
