--- BassClefStudio.Serialization/ISerializable.cs
@@ -1,14 +1,0 @@
-﻿using BassClefStudio.Serialization.Graphs;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BassClefStudio.Serialization
-{
-    /// <summary>
-    /// Indicates that the implementing types and anything derived from them supports <see cref="Node"/> creation.
-    /// </summary>
-    public interface ISerializable
-    {
-    }
-}
